CxGoScanner first commit
